empiezo con la entidad mascotas, metodo GET

+++ src/main/java/com/woofwk/woofworkout/domain/repository/MascotaRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.woofwk.woofworkout.models.Mascota;
 
 @Repository
-public interface MascotaRepository extends JpaRepository<Mascota, Long>{
+public interface MascotaRepository extends JpaRepository<Mascota, Integer>{
 
 }
